Update with Latest Version from Repo

+++ callUncrustify.sh
@@ -18,6 +18,9 @@ FILE_PATTERN=$DEF_PATTERN
 for i
 do
  case "$i" in
+  -c=*|--config=*)
+   CONFIGURATION_FILE="${i#*=}"
+   ;;
   -q|--quiet)
    SILENT=1
    ;;
@@ -32,10 +35,12 @@ do
    echo 'uncrustify on .c and .h C files in the current folder'
    echo ' only modifed files are rewritten'
    echo 'optional parameters:'
+   echo ' -c=... --config=...: defines configuration file to use'
    echo ' -q --quiet: does not display anything'
-   echo ' -r --recursive: recursively process of files'
+   echo ' -r --recursive: recursive process of files'
    echo ' -u=... --uncrustify=...: defines the uncrustify app to use'
    echo " \"<pattern>\": redefines file <pattern> instead \"${DEF_PATTERN}\" (\"${DEF_RPATTERN}\" if recursive)"
+   echo "              caveat: surrounding double-quotes (\"...\") matter much!"
    exit
    ;;
   *)
@@ -50,7 +55,7 @@ then
  echo "failure: cannot run ${UNCRUSTIFY_APP} app"
  echo "it is possible to define the app with option -u=<uncrustifyApp>"
  echo "note: under Windows, it seems necessary to specify the '.exe' extension"
- echo "        and shortcut '-u=.exe' on command line ensures this specification"
+ echo "        and shortcut '-u=.exe' on command line adds this extension"
  exit
 fi
 
@@ -62,22 +67,27 @@ fi
 
 # the actual job
 function doUncrust() {
- ${UNCRUSTIFY_APP} -f $1 -c ${CONFIGURATION_FILE} -o ${TEMP_FILE} >/dev/null 2>&1
- diff --brief $1 ${TEMP_FILE} >/dev/null 2>&1
- if [[ $? -eq 0 ]]
+ if [[ -f $1 ]]
  then
-  rm ${TEMP_FILE}
+  ${UNCRUSTIFY_APP} -f $1 -c ${CONFIGURATION_FILE} -o ${TEMP_FILE} >/dev/null 2>&1
+  diff --brief $1 ${TEMP_FILE} >/dev/null 2>&1
+  if [[ $? -eq 0 ]]
+  then
+   rm ${TEMP_FILE}
+  else
+   echo $1 UNCRUSTIFIED! 2>&1
+   mv ${TEMP_FILE} $1
+   NB_FILES_UNCRUSTIFIED=$((NB_FILES_UNCRUSTIFIED + 1))
+  fi
  else
-  echo $1 UNCRUSTIFIED! 2>&1
-  mv ${TEMP_FILE} $1
-  NB_FILES_UNCRUSTIFIED=$((NB_FILES_UNCRUSTIFIED + 1))
+  echo "$1 not found!" >/dev/null 2>&1
  fi
 }
 
 echo "Uncrustification started" 2>&1
 shopt -s globstar
 FILE_LIST="${FILE_PATTERN}"
- for FILE in $FILE_LIST
+for FILE in $FILE_LIST
  do
   doUncrust ${FILE}
  done
